+++ packages/queries/tests/Querying/Evaluators/ConditionEvaluatorTest.php
@@ -187,8 +187,8 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultsEqualWithTwo(): void
     {
-        $pathA = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'birth', 'day'));
-        $pathB = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'birth', 'month'));
+        $pathA = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['birth', 'day']));
+        $pathB = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['birth', 'month']));
         $birthMonthAndDaySimilar = new AllEqual($pathA, $pathB);
         $expected = [$this->authors['salinger']];
         $filteredAuthors = array_values($this->conditionEvaluator->filterArray($this->authors, $birthMonthAndDaySimilar));
@@ -197,9 +197,9 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultsEqualWithThree(): void
     {
-        $pathA = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'birth', 'day'));
-        $pathB = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'birth', 'month'));
-        $pathC = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'birth', 'year'));
+        $pathA = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['birth', 'day']));
+        $pathB = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['birth', 'month']));
+        $pathC = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['birth', 'year']));
         $birthMonthAndDaySimilar = new AllEqual($pathA, $pathB, $pathC);
         $filteredAuthors = $this->conditionEvaluator->filterArray($this->authors, $birthMonthAndDaySimilar);
         self::assertEquals([], $filteredAuthors);
@@ -207,7 +207,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultHasValue(): void
     {
-        $country = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'author', 'birth', 'country'));
+        $country = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['author', 'birth', 'country']));
         $england = new Value('England');
         $propertyHasValue = new AllEqual($country, $england);
         $expectedBooks = [$this->books['pickwickPapers'], $this->books['philosopherStone'], $this->books['deathlyHallows']];
@@ -217,7 +217,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultHasSize(): void
     {
-        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, 'books')));
+        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, ['books'])));
         $two = new Value(2);
         $hasSize = new AllEqual($size, $two);
         $expectedAuthors = [$this->authors['rowling']];
@@ -227,7 +227,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultHasDoubleSizeWithSum(): void
     {
-        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, 'books')));
+        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, ['books'])));
         $doubleSize = new Sum($size, $size);
         $two = new Value(4);
         $hasSize = new AllEqual($doubleSize, $two);
@@ -239,7 +239,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
     public function testFunctionLowerCase(): void
     {
         $constant = new Value('phen');
-        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'name'));
+        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['name']));
         $lower = new LowerCase($property);
         $functionCondition = new StringContains($lower, $constant, false);
         $expectedAuthors = [$this->authors['king']];
@@ -250,7 +250,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
     public function testFunctionLowerCaseFalse(): void
     {
         $constant = new Value('PHEN');
-        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'name'));
+        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['name']));
         $lower = new LowerCase($property);
         $functionCondition = new StringContains($lower, $constant, true);
         $filteredAuthors = $this->conditionEvaluator->filterArray($this->authors, $functionCondition);
@@ -260,7 +260,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
     public function testFunctionUpperCase(): void
     {
         $constant = new Value('PHEN');
-        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'name'));
+        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['name']));
         $upper = new UpperCase($property);
         $functionCondition = new StringContains($upper, $constant, false);
         $expectedAuthors = [$this->authors['king']];
@@ -271,7 +271,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
     public function testFunctionUpperCaseFalse(): void
     {
         $constant = new Value('phen');
-        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, 'name'));
+        $property = new Property(new PropertyPath(null, '', PropertyPath::UNPACK, ['name']));
         $upper = new UpperCase($property);
         $functionCondition = new StringContains($upper, $constant, true);
         $filteredAuthors = $this->conditionEvaluator->filterArray($this->authors, $functionCondition);
@@ -280,7 +280,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testFunctionResultHasDoubleSizeWithProduct(): void
     {
-        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, 'books')));
+        $size = new Size(new Property(new PropertyPath(null, '', PropertyPath::DIRECT, ['books'])));
         $doubleSize = new Product(new Value(2), $size);
         $two = new Value(4);
         $hasSize = new AllEqual($doubleSize, $two);
@@ -291,7 +291,7 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testCustomMemberCondition(): void
     {
-        $propertyPath = new PropertyPath(null, '', PropertyPath::DIRECT, 'books', 'title');
+        $propertyPath = new PropertyPath(null, '', PropertyPath::DIRECT, ['books', 'title']);
         $condition = new AllEqual(
             new Value(true),
             new AllEqual(
@@ -309,8 +309,8 @@ class ConditionEvaluatorTest extends ModelBasedTest
 
     public function testCustomMemberConditionWithSalt(): void
     {
-        $propertyPathA = new PropertyPath(null, 'a', PropertyPath::DIRECT, 'books', 'title');
-        $propertyPathB = new PropertyPath(null, 'b', PropertyPath::DIRECT, 'books', 'title');
+        $propertyPathA = new PropertyPath(null, 'a', PropertyPath::DIRECT, ['books', 'title']);
+        $propertyPathB = new PropertyPath(null, 'b', PropertyPath::DIRECT, ['books', 'title']);
         $condition = new AllEqual(
             new Value(true),
             new AllEqual(